Feat/testing

+++ BE/src/main/java/com/itjob/constant/OtpConstant.java
@@ -8,7 +8,7 @@ public final class OtpConstant {
     }
 
     public static final int OTP_LENGTH = 6;
-    public static final int OTP_MIN = (int) Math.pow(10, OTP_LENGTH - 1);
+    public static final int OTP_MIN = (int) Math.pow(10,(double) OTP_LENGTH - 1);
     public static final int OTP_MAX = (int) Math.pow(10, OTP_LENGTH) - 1;
     public static final Duration OTP_TTL = Duration.ofMinutes(5);
     public static final Duration ATTEMPT_TTL = Duration.ofMinutes(5);
